Studybox 443

+++ StudyBox/StudyBox.Core/ViewModels/CreateFlashcardViewModel.cs
@@ -16,7 +16,7 @@ using Windows.UI.Popups;
 
 namespace StudyBox.Core.ViewModels
 {
-  
+
 
     public class CreateFlashcardViewModel : ExtendedViewModelBase
     {
@@ -29,6 +29,7 @@ namespace StudyBox.Core.ViewModels
         private IInternetConnectionService _internetConnectionService;
 
         private string _question;
+        private bool _isPublic;
         private string _answer;
         private string _deckName;
         private string _headerMessage;
@@ -154,6 +155,22 @@ namespace StudyBox.Core.ViewModels
                     _answer = value;
                     RaisePropertyChanged("Answer");
                     RaisePropertyChanged("CurrentAnswerCharactersNumber");
+                }
+            }
+        }
+
+        public bool IsPublic
+        {
+            get
+            {
+                return _isPublic;
+            }
+            set
+            {
+                if (_isPublic != value)
+                {
+                    _isPublic = value;
+                    RaisePropertyChanged("IsPublic");
                 }
             }
         }
@@ -312,7 +329,7 @@ namespace StudyBox.Core.ViewModels
                             break;
 
                         case Mode.CreateFlashcardAndDeck:
-                            Deck createdDeck = await _restService.CreateDeck(new Deck("1", DeckName));
+                            Deck createdDeck = await _restService.CreateDeck(new Deck("1", DeckName, IsPublic));
                             Flashcard addedFlashcard = await _restService.CreateFlashcard(_flashcard, createdDeck.ID);
 
                             foreach (var tip in tips)
@@ -329,9 +346,13 @@ namespace StudyBox.Core.ViewModels
                             bool result = await _restService.UpdateFlashcard(_flashcard, deckId);
 
                             var oldTips = await _restService.GetTips(deckId, flashcardId);
-                            foreach (var oldTip in oldTips)
+
+                            if (oldTips!=null)
                             {
-                                await _restService.RemoveTip(deckId, flashcardId, oldTip.ID);
+                                foreach (var oldTip in oldTips)
+                                {
+                                    await _restService.RemoveTip(deckId, flashcardId, oldTip.ID);
+                                }
                             }
 
                             foreach (var tip in tips)
@@ -346,7 +367,24 @@ namespace StudyBox.Core.ViewModels
                 {
                     ShowErrorMessage(StringResources.GetString("OperationFailed"));
                 }
-                //TODO navigate
+
+                switch (_mode)
+                {
+                    case Mode.CreateFlashcardAndDeck:
+                        NavigationService.NavigateTo("DecksListView");
+                        Messenger.Default.Send<ReloadMessageToDecksList>(new ReloadMessageToDecksList(true));
+                        Messenger.Default.Send<MessageToMenuControl>(new MessageToMenuControl(true, false, false));
+                        break;
+
+                    default:
+                        NavigationService.NavigateTo("ManageDeckView");
+                        Messenger.Default.Send<MessageToMenuControl>(new MessageToMenuControl(true, false, false));
+                        Messenger.Default.Send<DataMessageToMenageDeck>(new DataMessageToMenageDeck(_deck));
+                        break;
+
+                }
+
+                
             }
         }
 
@@ -400,9 +438,12 @@ namespace StudyBox.Core.ViewModels
                     SubmitFormMessage = StringResources.GetString("Edit"); ;
 
                     var tips = await _restService.GetTips(deckInstance.ID, flashcardInstance.Id);
-                    foreach (var tip in tips)
+                    if (tips != null)
                     {
-                        TipsCollection.Add(new TipViewModel(tip.ID, tip.Prompt));
+                        foreach (var tip in tips)
+                        {
+                            TipsCollection.Add(new TipViewModel(tip.ID, tip.Prompt));
+                        }
                     }
 
 
@@ -415,12 +456,13 @@ namespace StudyBox.Core.ViewModels
                     Answer = "";
 
                     _flashcard = new Flashcard();
+                    HeaderMessage = StringResources.GetString("AddFlashcard");
+                    SubmitFormMessage = StringResources.GetString("AddFlashcard");
                 }
 
                 _deck = deckInstance;
 
-                HeaderMessage = StringResources.GetString("AddFlashcard");
-                SubmitFormMessage = StringResources.GetString("AddFlashcard");
+                
                 ShowDeckName = false;
 
                 TipsCollection = new ObservableCollection<TipViewModel>();
